Changed code in searchHandler.mjs to previous version
