+++ js/utils/sanitizeBeforeRender.mjs
